pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "litellm>=1.72.4",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["uv_build>=0.7.13,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "text_to_sql"
module-root = ""

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.11.13",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
exclude = [
    "playground"
]
